requirements.txt
biopython>=1.80
matplotlib>=3.7
numpy>=1.23

